pyproject.toml
[project]
name = "ai-therapist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "git-filter-repo>=2.47.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "ollama>=0.5.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "streamlit>=1.48.1",
    "twilio>=9.7.1",
    "unicorn>=2.1.3",
    "uvicorn>=0.35.0",
]
